requirements.txt
requests>=2.31.0
pyyaml>=6.0
azure-storage-blob>=12.19.0
scikit-learn>=1.4.0
joblib>=1.3.0
numpy>=1.26.0
pytest>=8.0.0
flask
